third_party/blink/web_tests/platform/generic/fast/forms/textarea/textarea-rows-cols-expected.txt

Test for edge cases of <textarea> rows and cols attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Default values
PASS defaultRows is 2
PASS defaultCols is 20
PASS defaultHeight > 0 is true
PASS defaultWidth > 0 is true
rows = 1
PASS textarea.rows is 1
PASS textarea.offsetHeight > 0 is true
PASS textarea.offsetHeight < defaultHeight is true
PASS textarea.offsetWidth is defaultWidth
rows = 2; should match default height
PASS textarea.rows is defaultRows
PASS textarea.offsetHeight is defaultHeight
PASS textarea.offsetWidth is defaultWidth
rows = 3
PASS textarea.rows is 3
PASS textarea.offsetHeight > defaultHeight is true
PASS textarea.offsetWidth is defaultWidth
rows; should be default height
PASS textarea.rows is defaultRows
PASS textarea.offsetHeight is defaultHeight
PASS textarea.offsetWidth is defaultWidth
rows = 0; should be default height
PASS textarea.rows is defaultRows
PASS textarea.offsetHeight is defaultHeight
PASS textarea.offsetWidth is defaultWidth
rows = -1; should be default height
PASS textarea.rows is defaultRows
PASS textarea.offsetHeight is defaultHeight
PASS textarea.offsetWidth is defaultWidth
rows = x; should be default height
PASS textarea.rows is defaultRows
PASS textarea.offsetHeight is defaultHeight
PASS textarea.offsetWidth is defaultWidth
cols = 1
PASS textarea.cols is 1
PASS textarea.offsetWidth > 0 is true
PASS textarea.offsetWidth < defaultWidth is true
PASS textarea.offsetHeight is defaultHeight
cols = 20; should match default width
PASS textarea.cols is defaultCols
PASS textarea.offsetWidth is defaultWidth
PASS textarea.offsetHeight is defaultHeight
cols = 40
PASS textarea.cols is 40
PASS textarea.offsetWidth > defaultWidth is true
PASS textarea.offsetHeight is defaultHeight
cols; should be default width
PASS textarea.cols is defaultCols
PASS textarea.offsetWidth is defaultWidth
PASS textarea.offsetHeight is defaultHeight
cols = 0; should be default width
PASS textarea.cols is defaultCols
PASS textarea.offsetWidth is defaultWidth
PASS textarea.offsetHeight is defaultHeight
cols = -1; should be default width
PASS textarea.cols is defaultCols
PASS textarea.offsetWidth is defaultWidth
PASS textarea.offsetHeight is defaultHeight
cols = x; should be default width
PASS textarea.cols is defaultCols
PASS textarea.offsetWidth is defaultWidth
PASS textarea.offsetHeight is defaultHeight
PASS successfullyParsed is true

TEST COMPLETE
